+++ config/test.exs
@@ -1,5 +1,7 @@
 import Config
 
+config :volt, :server, watch: false
+
 config :volt,
        :lint,
        Keyword.merge(Application.get_env(:volt, :lint, []),